No branch length on root
So far, the root was given a branch length of zero. However, this leads
to a syntax error when parsed by my older tools midRoot and
new2view. Even though these are being replaced by the new midRoot and
plotTree, for the sake of backward compatibility I leave out the branch
length on the root.